requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
deep-translator>=1.11.4
opencv-python>=4.8.0
numpy>=1.24.0